Use !empty() instead of isset() to check for array key in setuplib.php
In setuplib.php, initialise_fullme() runs the php function parse_url() on $CFG->wwwroot. Parse_url() returns an array containing all the components of the input URL. The problem is that if a URL component (such as port) is missing from the input URL, then it will be missing from the parse_url() return array. So we must check for the presence of each element. Since the rest of the function does this with !empty() then use !empty() rather than isset() to check for $wwwroot['scheme'].